--- client/src/components/Tooltip/TooltipComp.js
@@ -1,12 +1,0 @@
-import React from "react";
-import { Tooltip } from "antd";
-
-const TooltipComp = ({ placement, title, children }) => {
-	return (
-		<Tooltip placement={placement} title={title}>
-			{children}
-		</Tooltip>
-	);
-};
-
-export default TooltipComp;
